Liver_Diseases/requirements.txt
numpy==1.23.5
pandas==1.4.3
pydicom==2.3.1
Pillow==9.2.0
scikit-learn==1.0.2
openpyxl==3.0.10
tensorflow==2.10.0
matplotlib==3.5.2
seaborn==0.12.0 